chore: add dylib downloader and validator (#16)
First PR for #2.

This PR adds an abstraction for downloading & validating the dylib from a Coder server, and the network extension scaffolding.

It also adds a `TunnelHandle` type for owning the pair of pipes passed to the dylib, and the handle to the dylib itself.

You cannot create a unit test target that targets a System Extension. So, this PR extracts the portion of the network extension that we'd like to test into it's own Framework, `VPNLib`.

Of note is that `SwiftProtobuf` doesn't have a stable ABI (as it doesn't use [library evolution](https://www.swift.org/blog/library-evolution/)), so the Framework has the `Build libraries for distribution` setting disabled. This shouldn't effect anything. Exporting the `SwiftProtobuf` types should be fine provided we don't import `SwiftProtobuf` in to the `VPN` target as well.