fix: hold delegate reference to keep it from being deallocated (#26)
The problem we were having before was being caused by the system extension delegate being deallocated before the request responses came back.

I didn't initially realize this, but the [OSSystemRequest.delegate](https://developer.apple.com/documentation/systemextensions/ossystemextensionrequest/delegate) is a weak reference (you have to click down to the individual field view in the docs to learn this).

So, we keep a strong reference from the VPNService around while we are waiting for the extension to install.