chore: handle waking from device sleep (#50)
Depends on https://github.com/coder/coder/pull/16598.
Reverts #43.

Whilst everything seems to recover okay tailnet wise when waking from sleep currently, the tunnel will still miss workspace/peer updates during the sleep, causing the workspace state in the UI to be out of sync with reality. 

To handle this, we'll teardown the tunnel on sleep, and bring it back up on wake.

Fixing the issue in `coder/coder` also revealed that the error encountered when toggling the VPN on and off quickly was another symptom, and so this change reverts the code that prevents toggling the VPN on and off quickly, as it now works flawlessly.