chore(pkgbuild): delete existing app during preinstall, `spctl --assess` during postinstall (#112)
Relates to #83.

It looks like deleting the app does indeed kill the NE process, so we should do that during `preinstall`.

In case the XPC issue we've been seeing is due to a race between the app being opened, and Gatekeeper ingesting the notarization ticket, we'll also force Gatekeeper to read the ticket by running `spctl -a` on the app bundle, and the extension bundle. The latter always fails, but an attempt to read it can't hurt.