fix: start coder connect on launch after SE is installed (#113)
This is a fix for #108. Previously we would start the VPN immediately on launch if the config option was true. This starting of the VPN raced with any concurrent upgrades of the network extension, causing the VPN to be started with a VPN config belonging to the older network extension, and producing a consistent error message: 
![image](https://github.com/user-attachments/assets/a69932cb-4c86-4d45-8ab5-5843e255f395)

Instead, we should only start the VPN once we know that the system extension and VPN configuration are installed.