feat: add file sync daemon error handling to the UI (#122)
If file sync is working, but a session has errored, an icon will be displayed on the main menu.
e.g. for:
<img width="512" alt="image" src="https://github.com/user-attachments/assets/aac73c99-e318-44d3-9091-3f5d99239037" />
This icon & tooltip are displayed:
<img width="256" alt="image" src="https://github.com/user-attachments/assets/f4b7ba15-dca2-4819-aaa8-07e74e4e238d" />

If file sync is not working altogether, due to the daemon crashing, the same icon will be displayed with a different tooltip on hover:
<img width="254" alt="image" src="https://github.com/user-attachments/assets/efc87c1d-acac-4353-a3c9-c04908762d28" />

Once the config menu is opened, an alert is displayed, and the daemon log file is opened.
<img width="1354" alt="image" src="https://github.com/user-attachments/assets/98b44f6e-4584-4ad3-a237-1557ec5edab1" />

From there, the Daemon can be restarted, or the alert can be dismissed without restarting. The latter provides users an out if the daemon were to crash on launch repeatedly.