feat: add conflict descriptions and file sync context menu (#126)
Last QoL PR for now..

This adds buttons to the alt click context menu:
<img width="971" alt="Screenshot 2025-03-28 at 3 25 12 pm" src="https://github.com/user-attachments/assets/2477ee7d-2466-4fa5-9f7f-53a711ffdd64" />



And it adds a brief description of each conflict to the status tooltip:
<img width="405" alt="image" src="https://github.com/user-attachments/assets/e513faf1-414f-4612-902d-204b277a34b1" />

There's three cases for now. The first is just a basic file conflict, the second is if there's a type conflict (file, directory, symlink, etc), and the third is self-explanatory.

We'll need to come up with a proper design for how we show conflicts, so this implementation is just to not leave users in the dark if they run into any.