feat: add progress messages when creating sync sessions (#139)
This loading might take a minute on a poor connection, and there's currently no feedback indicating what's going on, so we can display the prompt messages in the meantime.

i.e. setting up a workspace with a fair bit of latency:

https://github.com/user-attachments/assets/4321fbf7-8be6-4d4b-aead-0581c609d668



This PR also contains a small refactor for the `Agent` `primaryHost`, removing all the subsequent nil checks as we know it exists on creation.