chore: improve performance of indeterminate spinner (#184)
A user reported a constant 10% CPU usage whilst the Cursor svg failed to load. It turns out unnecessarily tying a looping animation to some state in SwiftUI is a bad idea. If you want to render a looping animation that's not tied to some state, you should use the CoreAnimation framework. In this case, we use a `CABasicAnimation`.

We leave the determinate spinner unmodified, as it by definition must be tied to some SwiftUI state.

Before:
![before](https://github.com/user-attachments/assets/aadd00bd-d779-456d-9a2a-d72e24b085b1)
After:
![after](https://github.com/user-attachments/assets/ca788653-fbb2-469b-8bc8-2c0e5361945f)