fix: prompt for sign in when toggling coder connect on (#204)
I undid the work of #114 in #158, by disabling the VPN toggle when the network extension was unconfigured.
When signed out, the network extension is unconfigured.
This PR adds a special exception to make the VPN toggle always clickable when signed out, as it has special behaviour when signed out.
It also adds a proper regression test. A slightly different regression test existed, but it didn't account for cases where the VPN state is one that would normally disable the toggle.